pyproject.toml
[project]
name = "personalml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.3",
    "numpy==1.26.0",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "quart>=0.19.9",
    "scipy>=1.14.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
